# docs/swagger.yaml
basePath: /
definitions:
  api.createTodoRequest:
    properties:
      title:
        maxLength: 255
        type: string
    required:
    - title
    type: object
  api.getTodoAttachmentMetadataResponse:
    properties:
      attachmentId:
        type: integer
      filename:
        type: string
      todoId:
        type: integer
    type: object
  api.updateTodoRequestBody:
    properties:
      status:
        type: string
      title:
        maxLength: 255
        type: string
    type: object
  db.Todo:
    properties:
      createdAt:
        type: string
      fileCount:
        type: integer
      status:
        type: string
      title:
        type: string
      todoId:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: Bhavya Jain
  description: A todo management service API in go which supports attachments
  title: Todo API
  version: "1.0"
paths:
  /health:
    get:
      description: Returns server health
      produces:
      - text/plain
      responses:
        "200":
          description: OK
      tags:
      - health
  /todos:
    get:
      description: List todos based on page ID and page size
      parameters:
      - description: page ID
        in: query
        minimum: 1
        name: pageId
        required: true
        type: integer
      - description: page size
        in: query
        maximum: 10
        minimum: 5
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Todo'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: List todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Creates a todo with the specified title
      parameters:
      - description: Todo title
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/api.createTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Todo'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Creates a Todo
      tags:
      - todos
  /todos/{id}/attachments:
    get:
      consumes:
      - application/json
      description: Get attachment metadata for the corresponding todo
      parameters:
      - description: Todo ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.getTodoAttachmentMetadataResponse'
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get attachments metadata
      tags:
      - attachments
    post:
      consumes:
      - multipart/form-data
      description: Upload attachments for the corresponding todo
      parameters:
      - description: Todo ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - collectionFormat: csv
        description: attachments
        in: formData
        items:
          type: file
        name: attachments
        required: true
        type: array
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "413":
          description: Request Entity Too Large
        "415":
          description: Unsupported Media Type
        "500":
          description: Internal Server Error
      summary: Upload attachments
      tags:
      - attachments
  /todos/{id}/attachments/{attachmentId}:
    delete:
      consumes:
      - application/json
      description: Delete attachment for the corresponding todo
      parameters:
      - description: Todo ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: attachment ID
        in: path
        minimum: 1
        name: attachmentId
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "413":
          description: Request Entity Too Large
        "500":
          description: Internal Server Error
      summary: Delete attachment
      tags:
      - attachments
  /todos/{todoId}:
    delete:
      consumes:
      - application/json
      description: Delete todo by TodoID
      parameters:
      - description: Todo ID
        in: path
        minimum: 1
        name: todoId
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Deletes a Todo
      tags:
      - todos
    get:
      description: Get todo by TodoID
      parameters:
      - description: Todo ID
        in: path
        minimum: 1
        name: todoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Todo'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Returns a Todo
      tags:
      - todos
    patch:
      consumes:
      - application/json
      description: Updates the todo title/status
      parameters:
      - description: Todo ID
        in: path
        minimum: 1
        name: todoId
        required: true
        type: integer
      - description: Todo title/status
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/api.updateTodoRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Todo'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Updated the todo title/status
      tags:
      - todos
  /todos/{todoId}/attachments/{attachmentId}:
    get:
      consumes:
      - application/json
      description: Get attachment for the corresponding todo
      parameters:
      - description: Todo ID
        in: path
        minimum: 1
        name: todoId
        required: true
        type: integer
      - description: attachment ID
        in: path
        minimum: 1
        name: attachmentId
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get attachments
      tags:
      - attachments
swagger: "2.0"